Project Description:
This application will be used to make a fake call to the device. This application is generally essential for emergency situations. This can be good for situations where the users are alone and in danger. This app will help them pretend like they’re on call and walk out of the path.

To develop this application, Android Studio is required. For programming, Java and XML will be required in Android Studio. For fake calling, you need to get access to the CallManager package in Java.